Create a Component that accepts an array of strings as input and it will display all array items as an unordered-list.
Create a custom Directive called "upper" that once applied to an element like 
<li> it's going to convert the content of the element to uppercase.

Create a custom Directive called "myvisibility" that accepts a boolean input
which is going to show or hide the element.

Create a custom Directive called "mycolor" when clicked it will change its color
from an array of colors and emit the color value to parent component
to show the value of the current selected color of the clicked component.